inttypes

+++ crypto/CryptoAuth.c
@@ -567,7 +567,7 @@ static uint8_t sendMessage(struct Message* message, struct Interface* interface)
     uint64_t nowSecs = Time_currentTimeSeconds(wrapper->context->eventBase);
     if (nowSecs - wrapper->context->resetAfterInactivitySeconds > wrapper->timeOfLastPacket) {
         Log_debug(wrapper->context->logger, "No traffic in [%d] seconds, resetting connection.",
-                  nowSecs - wrapper->timeOfLastPacket);
+                  (int) (nowSecs - wrapper->timeOfLastPacket));
         wrapper->timeOfLastPacket = nowSecs;
         CryptoAuth_reset(interface);
         return encryptHandshake(message, wrapper);
@@ -575,7 +575,7 @@ static uint8_t sendMessage(struct Message* message, struct Interface* interface)
 
     // debugging...
     Log_debug(wrapper->context->logger, "No incoming traffic in [%d] seconds",
-              nowSecs - wrapper->timeOfLastPacket);
+              (int) (nowSecs - wrapper->timeOfLastPacket));
 
     #ifdef Log_DEBUG
         Assert_true(!((uintptr_t)message->bytes % 4) || !"alignment fault");
